Create Repeated String
Lilah has a string,s , of lowercase English letters that she repeated infinitely many times.

Given an integer,n , find and print the number of letter a's in the first n letters of Lilah's infinite string.

For example, if the string s='abcac' and ,n=10 the substring we consider is abcacabcac, the first 10 characters of her infinite string. There are 4 occurrences of a in the substring.10

Sample Input 0

aba
10
Sample Output 0

7
Explanation 0
The first n=10 letters of the infinite string are abaabaabaa. Because there are 7 a's, we print 7 on a new line.

Sample Input 1

a
1000000000000
Sample Output 1

1000000000000