Two Strings
Given two strings, determine if they share a common substring. A substring may be as small as one character.

For example, the words "a", "and", "art" share the common substring a. The words "be" and "cat" do not share a substring.

Sample Input

2
hello
world
hi
world
Sample Output

YES
NO